// src/Links.enum.ts
export enum Social {
  Github = 'https://github.com/meskal1',
  Codewars = 'https://www.codewars.com/users/meskal1',
  Telegram = 'https://t.me/DaniilKalach',
  Linkedin = 'https://www.linkedin.com/in/daniilkalach/',
}

export enum DemoProject {
  Portfolio = 'https://meskal1.github.io/portfolio/',
  LottieAnimation = 'https://lottiefiles.com/122697-get-skills-from-the-box',
  SocialNetwork = 'https://meskal1.github.io/social_network/',
  Todolist = 'https://meskal1.github.io/my_todo_list/',
  Counter = 'https://meskal1.github.io/Counter',
  DrumKit = 'https://meskal1.github.io/drum_kit/',
  GuessTheNumber = 'https://meskal1.github.io/guess_my_number/',
  PigGame = 'https://meskal1.github.io/pig_game/',
  CarAnimation = 'https://meskal1.github.io/CSS_car_animation/',
  LandingPractice = 'https://meskal1.github.io/landing_practice/',
}

export enum SourceProject {
  Portfolio = 'https://github.com/meskal1/portfolio',
  LottieAnimation = 'https://lottiefiles.com/122697-get-skills-from-the-box',
  SocialNetwork = 'https://github.com/meskal1/social_network',
  Todolist = 'https://github.com/meskal1/my_todo_list',
  Counter = 'https://github.com/meskal1/Counter',
  DrumKit = 'https://github.com/meskal1/drum_kit',
  GuessTheNumber = 'https://github.com/meskal1/guess_my_number',
  PigGame = 'https://github.com/meskal1/pig_game',
  CarAnimation = 'https://github.com/meskal1/CSS_car_animation',
  LandingPractice = 'https://github.com/meskal1/landing_practice',
}

export enum Image {
  Portfolio = 'https://github.com/meskal1/portfolio/raw/main/PreviewPortfolio.jpg',
  LottieAnimation = 'https://github.com/meskal1/portfolio/raw/main/LottiePreview.jpg',
  SocialNetwork = 'https://github.com/meskal1/social_network/raw/main/Preview.jpg',
  Todolist = 'https://github.com/meskal1/my_todo_list/raw/main/Preview.jpg',
  Counter = 'https://github.com/meskal1/Counter/raw/main/Preview.jpg',
  DrumKit = 'https://github.com/meskal1/drum_kit/raw/main/PreviewDrumKit.jpg',
  GuessTheNumber = 'https://github.com/meskal1/guess_my_number/raw/main/Preview.jpg',
  PigGame = 'https://github.com/meskal1/pig_game/raw/main/Preview.jpg',
  CarAnimation = 'https://github.com/meskal1/CSS_car_animation/raw/main/PreviewCar.jpg',
  LandingPractice = 'https://github.com/meskal1/landing_practice/raw/main/PreviewLanding.jpg',
}